legalgpt-local/requirements.txt
Flask==3.0.3
Flask-SQLAlchemy==3.1.1
python-dotenv==1.0.1
psycopg[binary]==3.2.13
pgvector==0.3.5
requests==2.32.3
Werkzeug==3.0.3
pypdf==4.3.1
python-docx==1.1.2
